requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
ndg-httpsclient==0.4.2
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==17.0.0
python-dotenv==0.6.4
requests==2.18.1
six==1.10.0
urllib3==1.21.1
